Refactor http request processing. This is the first cut.
