+++ modules/engine/lib/engine.js
@@ -32,7 +32,7 @@ var configLoader = require('./engine/config.js'),
     insert = require('./engine/insert.js'),
     jsonfill = require('./engine/jsonfill.js'),
     eventTypes = require('./engine/event-types.js'),
-    httpRequest = require('./engine/http.request.js'),
+    httpRequest = require('./engine/http/table.js'),
     LogEmitter = require('./engine/log-emitter.js'),
     winston = require('winston'),
     compiler = require('ql.io-compiler'),
