+++ modules/engine/lib/xformers/csv.js
@@ -39,4 +39,6 @@ exports.toJson = function(data, respCb, errorCb, headers) {
     catch(error) {
         return errorCb(error);
     }
-}
+};
+
+exports.accept = 'text/csv';
