+++ modules/engine/lib/xformers/json.js
@@ -23,4 +23,6 @@ exports.toJson = function(data, respCb, errorCb) {
     catch(error){
         return errorCb(error);
     }
-}
+};
+
+exports.accept = 'application/json';
