+++ modules/engine/lib/xformers/xml.js
@@ -25,4 +25,6 @@ exports.toJson = function(data, respCb, errorCb) {
     catch(error) {
         return errorCb(error);
     }
-}
+};
+
+exports.accept = 'application/xml';
